# swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/seguimiento/{id}": {
            "get": {
                "operationId": "SeguimientoController_get",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento"
                ]
            },
            "put": {
                "operationId": "SeguimientoController_put",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeguimientoDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento"
                ]
            },
            "delete": {
                "operationId": "SeguimientoController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "seguimiento"
                ]
            }
        },
        "/tipo-seguimiento": {
            "post": {
                "operationId": "TipoSeguimientoController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoSeguimientoDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "tipo-seguimiento"
                ]
            },
            "get": {
                "operationId": "TipoSeguimientoController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "query - Filter. e.g. col1:v1,col2:v2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "tipo-seguimiento"
                ]
            }
        },
        "/tipo-seguimiento/{id}": {
            "get": {
                "operationId": "TipoSeguimientoController_get",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "tipo-seguimiento"
                ]
            },
            "put": {
                "operationId": "TipoSeguimientoController_put",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoSeguimientoDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "tipo-seguimiento"
                ]
            },
            "delete": {
                "operationId": "TipoSeguimientoController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "tipo-seguimiento"
                ]
            }
        },
...
    },
    "info": {
        "title": "plan_crud",
        "description": "API CRUD para el registro de planes para el cliente de planeacion",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "SeguimientoDto": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "plan_id": {
                        "type": "string"
                    },
                    "dato": {
                        "type": "string"
                    },
                    "tipo_seguimiento_id": {
                        "type": "string"
                    },
                    "estado_seguimiento_id": {
                        "type": "string"
                    },
                    "periodo_seguimiento_id": {
                        "type": "string"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "fecha_inicio": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_fin": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_creacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_modificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "nombre",
                    "descripcion",
                    "plan_id",
                    "dato",
                    "tipo_seguimiento_id",
                    "estado_seguimiento_id",
                    "periodo_seguimiento_id",
                    "activo",
                    "fecha_inicio",
                    "fecha_fin",
                    "fecha_creacion",
                    "fecha_modificacion"
                ]
            },
            "TipoSeguimientoDto": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "codigo_abreviacion": {
                        "type": "string"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "fecha_creacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_modificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "nombre",
                    "descripcion",
                    "codigo_abreviacion",
                    "activo",
                    "fecha_creacion",
                    "fecha_modificacion"
                ]
            },
...
        }
    }
}